Update README.md
I may not be able to issue a wtfpl licence to it since other contributors have attached an MIT licence. Unsure though, anyway do what the fuck you want with my stuff but please understand that that licence only goes as far as my personal work in this repository. Other types of licencing may apply for work I have derived from and are integral parts of this repository. So I added these facts to the readme to avoid getting in any kind of trouble.